ipcache: Document remote-node identity assumption
As described in the comment added by this commit message, this logic is
making assumptions about the way that identities for remote nodes are
allocated, which hold true today but may not hold true in the future. If
we do end up providing a way to associate additional remote-node labels
with IPs corresponding to remote nodes, hopefully the future contributor
will be able to reason about the impact and prevent weird bugs /
interactions using this new comment.

Signed-off-by: Joe Stringer <joe@cilium.io>